Override default delete behavior to mark as deleted.
